Dynamic lookup of elm executables
As Elm 0.19 gets more popular, it's becoming more difficult to have
only a single installation of Elm on the system. Additionally,
homebrew doesn't help here and is actually a problem as it maintains
the latest version only (points go to Nixpkg here). As a workaround
for now, look in the `node_modules` directory to see if elm is
installed as a local dependency and use that version.